Removed utils/src/config.rs, put content into lib.rs.
